basic services

+++ Domain/Common/IBaseRepository.cs
@@ -8,7 +8,7 @@ namespace Domain.Common
 {
     public interface IBaseRepository<T> where T : class
     {
-        public Task<T> GetByIdAsync(int id);
+        public Task<T> GetByIdAsync(long id);
         public Task<ICollection<T>> GetAllSync();
         public Task<T> AddAsync(T entity);
         public Task<T> UpdateAsync(T entity);
